@@ -15,8 +15,8 @@
   },
   "devDependencies": {
     "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@testing-library/svelte": "^5.2.7",
     "@types/gapi.client.drive-v3": "^0.0.4",
     "@typescript-eslint/eslint-plugin": "^5.45.0",
@@ -25,21 +25,21 @@
     "autoprefixer": "^10.4.14",
     "eslint": "^8.28.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-svelte": "^2.30.0",
-    "flowbite-svelte": "^0.44.5",
-    "flowbite-svelte-icons": "^0.4.2",
+    "eslint-plugin-svelte": "^2.45.1",
+    "flowbite-svelte": "^0.46.0",
+    "flowbite-svelte-icons": "^1.0.0",
     "jsdom": "^26.0.0",
     "postcss": "^8.4.24",
     "postcss-load-config": "^4.0.1",
-    "prettier": "^2.8.0",
-    "prettier-plugin-svelte": "^2.10.1",
+    "prettier": "^3.1.0",
+    "prettier-plugin-svelte": "^3.2.6",
     "sass": "^1.64.2",
-    "svelte": "^4.0.5",
-    "svelte-check": "^3.4.3",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tailwindcss": "^3.3.2",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.0",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vitest": "^3.0.7"
   },
   "type": "module",
