@@ -2,7 +2,7 @@
   import Settings from '$lib/components/Settings/Settings.svelte';
   import { UserSettingsSolid } from 'flowbite-svelte-icons';
 
-  let settingsHidden = true;
+  let settingsHidden = $state(true);
 
   function openSettings() {
     settingsHidden = false;
@@ -10,7 +10,7 @@
 </script>
 
 <button
-  on:click={openSettings}
+  onclick={openSettings}
   class="hover:text-primary-700 hover:mix-blend-normal mix-blend-difference z-10 fixed opacity-50 hover:opacity-100 right-10 top-5 p-10 m-[-2.5rem]"
 >
   <UserSettingsSolid />
