@@ -1,7 +1,12 @@
 <script lang="ts">
   import { initPanzoom } from './util';
+  interface Props {
+    children?: import('svelte').Snippet;
+  }
+
+  let { children }: Props = $props();
 </script>
 
 <div use:initPanzoom>
-  <slot />
+  {@render children?.()}
 </div>
