@@ -1,4 +1,9 @@
-<slot />
+<script>
+  /** @type {{children?: import('svelte').Snippet}} */
+  let { children } = $props();
+</script>
+
+{@render children?.()}
 
 <style>
   :global(body.reader) {
