@@ -12,14 +12,14 @@
   const volume = $page.url.searchParams.get('volume');
   const url = `${BASE_URL}/${manga}/${volume}`;
 
-  let message = 'Loading...';
+  let message = $state('Loading...');
 
   let files: File[] = [];
 
-  let completed = 0;
-  let max = 0;
+  let completed = $state(0);
+  let max = $state(0);
 
-  $: progress = Math.floor((completed / max) * 100).toString();
+  let progress = $derived(Math.floor((completed / max) * 100).toString());
 
   async function onImport() {
     const mokuroRes = await fetch(url + '.mokuro', { cache: 'no-store' });
