@@ -4,8 +4,8 @@
   import ReaderToggles from './ReaderToggles.svelte';
   import { settings, updateSetting } from '$lib/settings';
 
-  let swipeThresholdValue = $settings.swipeThreshold;
-  let edgeButtonWidthValue = $settings.edgeButtonWidth;
+  let swipeThresholdValue = $state($settings.swipeThreshold);
+  let edgeButtonWidthValue = $state($settings.edgeButtonWidth);
   function onSwipeChange() {
     updateSetting('swipeThreshold', swipeThresholdValue);
   }
@@ -16,7 +16,9 @@
 </script>
 
 <AccordionItem>
-  <span slot="header">Reader</span>
+  {#snippet header()}
+    <span >Reader</span>
+  {/snippet}
   <div class="flex flex-col gap-5">
     <ReaderSelects />
     <hr class="border-gray-100 opacity-10" />
